windows: Avoid warning about format string type mismatch
GUID member Data1 is unsigned long, format string expects unsigned int.

+++ src/brickd/usb_windows.c
@@ -75,7 +75,7 @@ bool usb_check_hotplug_event(USBHotplugType type, GUID *guid, const char *name) 
 
 	snprintf(guid_str, sizeof(guid_str),
 	         "{%08X-%04X-%04X-%02X%02X-%02X%02X%02X%02X%02X%02X}",
-	         guid->Data1, guid->Data2, guid->Data3,
+	         (uint32_t)guid->Data1, guid->Data2, guid->Data3,
 	         guid->Data4[0], guid->Data4[1], guid->Data4[2], guid->Data4[3],
 	         guid->Data4[4], guid->Data4[5], guid->Data4[6], guid->Data4[7]);
